Fixed an issue with the supression of occupations for the members

+++ server/api/OccupationsAPI.js
@@ -1,7 +1,8 @@
 //----------------Dependencies-------------//
 let Occupations     = require("../models/OccupationsMD.js"),
     Api             = new Object(),
-    Utility         = require("../utils/utility.js");
+    Utility         = require("../utils/utility.js"),
+    Members         = require("../models/MembersMD.js");
 
 //--------------Model-------------//
 
@@ -45,14 +46,37 @@ Api.UpdateOccupations = function(req, res)
 
 Api.DeleteOccupations = function(req , res)
 {
-    Occupations.findByIdAndRemove(req.params.id)
-        .then((occupation) =>{
-            Utility.GenerateResponse(true, res, occupation);
-        })
-        .catch((err) => {
-            Utility.GenerateResponse(false, res, err);
-            Utility.WriteInLog("error", err);
-        });
+    let canDelete = true;
+
+    Members.find({Occupation : req.params.id})
+            .then((members) =>{
+                if(members.length > 0){
+                    canDelete = false;
+
+                    let Poste = {
+                        kind: "dependencies",
+                        message: "Le poste est utilisé par un membre",
+                        name: "Poste",
+                        path: "Poste"
+                    }
+
+                    Utility.GenerateResponse(false, res, {errors : {Poste : Poste}});
+                }
+            })
+            .then(() =>{
+                if(canDelete)
+                {
+                    Occupations.findByIdAndRemove(req.params.id)
+                        .then((occupation) =>{
+                            Utility.GenerateResponse(true, res, occupation);
+                        })
+                        .catch((err) => {
+                            Utility.GenerateResponse(false, res, err);
+                            Utility.WriteInLog("error", err);
+                        });
+                }
+            });
+
 }
 
 //----------------Module Exports-------------//
